+++ packages/main/src/RPA/Excel/Application.py
@@ -11,6 +11,10 @@ if platform.system() == "Windows":
     import win32api
     import win32com.client
     import pywintypes
+else:
+    logging.getLogger(__name__).warning(
+        "RPA.Excel.Application library works only on Windows platform"
+    )
 
 
 def _to_unsigned(val):